+++ client/my-sites/site-settings/settings-discussion/controller.js
@@ -7,7 +7,7 @@ import React from 'react';
  * Internal dependencies
  */
 import { renderWithReduxStore } from 'lib/react-helpers';
-import DiscussionMain from 'my-sites/site-settings/discussion/main';
+import DiscussionMain from 'my-sites/site-settings/settings-discussion/main';
 
 export default {
 	discussion( context ) {
